Extract PreparedCode
